% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annofuns.R
\name{af}
\alias{af}
\title{Internal constructor for annotation functions}
\usage{
af(call, env = parent.frame())
}
\arguments{
\item{env}{}
}
\value{

}
\description{
Internal constructor for annotation functions
}
