% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annofuns.R
\name{af_call_highlight}
\alias{af_call_highlight}
\title{Title}
\usage{
af_call_highlight(color = NA, fill = NA, geomfun = ggplot2::geom_point, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
