% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annofuns.R
\name{af_call_label}
\alias{af_call_label}
\title{Title}
\usage{
af_call_label(label, geomfun, ...)
}
\arguments{
\item{...}{Arguments to pass on to ggrepel::geom_text/label_repel function.}

\item{bordered}{If TRUE, use ggrepel::geom_label_repel; otherwise use ggrepel::geom_text_repel.}
}
\value{

}
\description{
Title
}
