% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ap_shiny.R
\name{APServer}
\alias{APServer}
\title{Functions for adding 'annoplot' objects to a Shiny app}
\usage{
APServer(id, dplist, goreactive, points, brush = T, scale = 1.5, hover = F)
}
\arguments{
\item{hover}{}
}
\description{
Functions for adding 'annoplot' objects to a Shiny app
}
\seealso{
Shiny modules for AnnoPlot objects
\code{\link{multiAP}()}
}
\concept{shinymod}
