% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiap_shiny.R
\name{multiAP}
\alias{multiAP}
\title{Module functions for displaying multiple 'annoplot' objects in one UI block.}
\usage{
multiAP(id)
}
\arguments{
\item{input}{}

\item{...}{}
}
\description{
These are called internally by the
}
\seealso{
Shiny modules for AnnoPlot objects
\code{\link{APServer}()}
}
\concept{shinymod}
